refactor(connect): move account method definitions to corresponding folder
