+++ packages/connect-common/src/types/api/account/getAccountInfo.ts
@@ -1,8 +1,8 @@
 import type { BlockchainLinkParams } from '@trezor/blockchain-link';
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { AccountInfo, DiscoveryAccountType } from '../account';
-import type { BundledParams, Params, Response } from '../params';
+import type { AccountInfo, DiscoveryAccountType } from '../../account';
+import type { BundledParams, Params, Response } from '../../params';
 
 export interface GetAccountInfo extends Omit<BlockchainLinkParams<'getAccountInfo'>, 'descriptor'> {
     coin: string;
