+++ packages/connect-common/src/types/api/account/getCoinInfo.ts
@@ -1,0 +1,4 @@
+import type { CoinInfo } from '../../coinInfo';
+import type { CommonParams, Response } from '../../params';
+
+export declare function getCoinInfo(params: CommonParams & { coin: string }): Response<CoinInfo>;
