+++ packages/connect-common/src/types/api/account/signMessage.ts
@@ -6,7 +6,7 @@
 
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
-import type { SignMessage } from './bitcoin/common';
+import type { Params, Response } from '../../params';
+import type { SignMessage } from './../bitcoin/common';
 
 export declare function signMessage(params: Params<SignMessage>): Response<PROTO.MessageSignature>;
