+++ packages/connect-common/src/types/api/account/verifyMessage.ts
@@ -5,7 +5,7 @@
 
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
-import type { VerifyMessage } from './bitcoin/common';
+import type { Params, Response } from '../../params';
+import type { VerifyMessage } from './../bitcoin/common';
 
 export declare function verifyMessage(params: Params<VerifyMessage>): Response<PROTO.Success>;
