--- packages/connect-common/src/types/api/getCoinInfo.ts
@@ -1,4 +1,0 @@
-import type { CoinInfo } from '../coinInfo';
-import type { CommonParams, Response } from '../params';
-
-export declare function getCoinInfo(params: CommonParams & { coin: string }): Response<CoinInfo>;
